requirements.txt
beautifulsoup4==4.4.1
click==6.6
cookies==2.2.1
Flask==0.11.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PySocks==1.5.7
python-forecastio==1.3.4
pytz==2016.4
requests==2.10.0
responses==0.5.1
six==1.10.0
twilio==5.4.0
Werkzeug==0.11.10
wikipedia==1.4.0
